feat: organizar o estado dos projetos
