+++ alura-tracker/src/components/FormularioTarefa.vue
@@ -85,7 +85,7 @@ export default defineComponent({
     const store = useStore();
     return {
       store,
-      projetos: computed(() => store.state.projetos)
+      projetos: computed(() => store.state.projeto.projetos)
     }
   }
 });
