+++ alura-tracker/src/store/index.ts
@@ -1,44 +1,27 @@
+import http from "@/http";
 import { InjectionKey } from "vue";
 import { createStore, Store, useStore as vuexUserStore } from "vuex";
-import IProjeto from "@/interfaces/IProjetos";
 import { INotificacao } from "@/interfaces/INotificacao";
-import { ADICIONA_PROJETO, ADICIONA_TAREFA, ALTERA_PROJETO, ALTERA_TAREFA, DEFINIR_PROJETOS, DEFINIR_TAREFAS, EXCLUI_PROJETO, NOTIFICAR } from "./type-mutations";
-import { ALTERAR_PROJETO, ALTERAR_TAREFA, CADASTRAR_PROJETO, CADASTRAR_TAREFA, OBTER_PROJETOS, OBTER_TAREFAS, REMOVER_PROJETO } from "./type-actions";
-import http from "@/http";
+import { ADICIONA_TAREFA, ALTERA_TAREFA, DEFINIR_TAREFAS, NOTIFICAR } from "./type-mutations";
+import { ALTERAR_TAREFA, CADASTRAR_TAREFA, OBTER_TAREFAS } from "./type-actions";
+import { EstadoProjeto, moduloProjeto } from "./modules/projeto";
 import ITarefa from "@/interfaces/ITarefa";
 
-interface Estado {
+export interface Estado {
     tarefas: ITarefa[],
-    projetos: IProjeto[],
     notificacoes: INotificacao[],
+    projeto: EstadoProjeto,
 }
 
 export const key: InjectionKey<Store<Estado>> = Symbol();
 
 export const store = createStore<Estado>({
     state: {
-        projetos: [],
         notificacoes: [],
         tarefas: [],
+        projeto: { projetos: [] },
     },
     mutations: {
-        [ADICIONA_PROJETO](state, nomeProjeto: string) {
-            const projeto: IProjeto = {
-                id: new Date().toISOString(),
-                nome: nomeProjeto,
-            };
-            state.projetos.push(projeto);
-        },
-        [ALTERA_PROJETO](state, projeto: IProjeto) {
-            const index = state.projetos.findIndex(proj => proj.id == projeto.id);
-            state.projetos[index] = projeto;
-        },
-        [EXCLUI_PROJETO](state, id: string) {
-            state.projetos = state.projetos.filter(proj => proj.id != id);
-        },
-        [DEFINIR_PROJETOS](state, projetos: IProjeto[]) {
-            state.projetos = projetos;
-        },
         [DEFINIR_TAREFAS](state, tarefas: ITarefa[]) {
             state.tarefas = tarefas;
         },
@@ -62,20 +45,6 @@ export const store = createStore<Estado>({
         },
     },
     actions: {
-        [OBTER_PROJETOS]({ commit }) {
-            http.get("projetos")
-                .then(resposta => commit(DEFINIR_PROJETOS, resposta.data));
-        },
-        [CADASTRAR_PROJETO](contexto, nomeDoProjeto: string) {
-            return http.post("/projetos", { nome: nomeDoProjeto });
-        },
-        [ALTERAR_PROJETO](contexto, projeto: IProjeto) {
-            return http.put(`/projetos/${projeto.id}`, projeto);
-        },
-        [REMOVER_PROJETO]({ commit }, id: string) {
-            return http.delete(`/projetos/${id}`)
-                .then(() => commit(EXCLUI_PROJETO, id));
-        },
         [OBTER_TAREFAS]({ commit }) {
             http.get("tarefas")
                 .then(resposta => commit(DEFINIR_TAREFAS, resposta.data));
@@ -89,6 +58,9 @@ export const store = createStore<Estado>({
                 .then(tarefa => commit(ALTERA_TAREFA, tarefa));
         },
     },
+    modules: {
+        projeto: moduloProjeto,
+    }
 });
 
 export function useStore(): Store<Estado> {
