+++ alura-tracker/src/views/Projetos/Lista.vue
@@ -16,7 +16,7 @@ export default defineComponent({
 
     return {
       store,
-      projetos: computed(() => store.state.projetos),
+      projetos: computed(() => store.state.projeto.projetos),
     }
   },
 });
